Refactor calibration modal into custom hook
